pyproject.toml
[tool.poetry]
name = "dine-server"
version = "0.1.0"
description = ""
authors = ["phillychi3 <phillychi3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.51.1"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
googlemaps = "^4.10.0"
geopy = "^2.4.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
